Revert "Templated RTree and extended its benchmarks"
This reverts commit c810caaaa7f9eb43384f86af869a1aede75ff6a0.

+++ tools/test_utils/include/TestUtils/MakeMeshes.hpp
@@ -57,8 +57,8 @@ std::shared_ptr<meshkernel::Mesh2D> MakeRectangularMeshForTesting(
     meshkernel::Point const& origin = {0.0, 0.0});
 
 std::shared_ptr<meshkernel::Mesh2D> MakeRectangularMeshForTesting(
-    size_t n,
-    size_t m,
+    int n,
+    int m,
     double delta,
     meshkernel::Projection projection,
     meshkernel::Point const& origin = {0.0, 0.0});
