+++ tools/test_utils/src/MakeMeshes.cpp
@@ -220,8 +220,8 @@ std::shared_ptr<meshkernel::Mesh2D> MakeRectangularMeshForTesting(
 }
 
 std::shared_ptr<meshkernel::Mesh2D> MakeRectangularMeshForTesting(
-    size_t n,
-    size_t m,
+    int n,
+    int m,
     double delta,
     meshkernel::Projection projection,
     meshkernel::Point const& origin)
